apps/agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.55.0
httpx>=0.27.0
pydantic>=2.0.0
mem0ai>=1.0.0
python-dotenv>=1.0.0
boto3>=1.35.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Optional: Google Calendar API for meeting scheduling
# Install only if you want to enable the schedule_meeting tool
# google-auth>=2.23.0
# google-auth-oauthlib>=1.1.0
# google-api-python-client>=2.100.0
